# api/src/openapi/api.yaml
openapi: 3.0.3
info:
  title: Condat Building Project - OpenAPI 3.0
  description: API for Condat Building
  version: 1.0.0
...
paths:
...
  /api/average:
    get:
      tags:
        - person-controller
      operationId: average
      summary: Average number of people in the building
      parameters:
        - name: start
          in: query
          required: true
          schema:
            type: string
            format: date-time
            example: "2018-03-20T09:12:28Z"
        - name: end
          in: query
          required: true
          schema:
            type: string
            format: date-time
            example: "2018-03-20T09:12:28Z"
      responses:
        '200':
          description: Average number of people in the building
          content:
            application/json:
              schema:
                type: string
                example: "Average: 2.5"
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                type: string
                example: "Invalid time interval"